refactor: entities/pick 폴더 제거
