+++ frontend/techpick/src/apis/pick/getPick/useGetPickQuery.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { getPick } from './getPick';
-import type { GetPickResponseType } from '../../type';
+import type { GetPickResponseType } from '../pickApi.type';
 
 // todo: suspense query로 바꾸기.
 export const useGetPickQuery = (pickId: number) => {
