+++ frontend/techpick/src/apis/pick/pickApi.type.ts
