+++ frontend/techpick/src/components/PickCard/pickCard.css.ts
