+++ frontend/techpick/src/components/PickCardGridLayout/pickCardGridLayout.css.ts
