+++ frontend/techpick/src/components/index.ts
@@ -2,3 +2,5 @@ export * from './Text';
 export * from './Button';
 export * from './Gap';
 export { DeferredComponent } from './DeferredComponent';
+export { PickCard } from './PickCard/PickCard';
+export { PickCardGridLayout } from './PickCardGridLayout/PickCardGridLayout';
