--- frontend/techpick/src/entities/pick/index.ts
@@ -1,3 +1,0 @@
-export { PickCard } from './ui/PickCard';
-export { PickCardGridLayout } from './ui/PickCardGridLayout';
-export * from './api';
